Fix OpenCV install.

+++ libs/install_opencv.sh
@@ -60,7 +60,9 @@ sudo apt-get install -y unzip wget
 wget https://github.com/opencv/opencv/archive/${OPENCV_VERSION}.zip
 unzip ${OPENCV_VERSION}.zip
 rm ${OPENCV_VERSION}.zip
-mv opencv-${OPENCV_VERSION} OpenCV
+if [! -e OpenCV]; then
+    mv opencv-${OPENCV_VERSION} OpenCV
+fi
 cd OpenCV
 mkdir build
 cd build
